✨ feat: 앱 초기화 로직을 커스텀 훅으로 분리
- src/basic/hooks/useAppInitialization.js 파일 생성
- initializeAppState, setupStateSubscription 함수 분리 및 훅으로 구성
- main.basic.js에서 해당 훅을 import하여 초기화 로직 간소화
- PRODUCT_LIST import도 useAppInitialization 내부로 이동
